Adds --tags to the `git fetch` Command
The `git fetch` command only pulls down tags that are reachable by the
current branch. There are cases where not all tags are reachable so not
all tags are fetched unless --tags is passed with fetch.

This commit adds --tags to the `git fetch` command to make sure the
`git checkout` command used by shipjs-deploy always succeeds if the
requested tag exists.

Updates tests to account for --tags.